+++ packages/effect/src/internal/config.ts
@@ -10,12 +10,14 @@ import * as HashSet from "../HashSet.js"
 import type * as LogLevel from "../LogLevel.js"
 import * as Option from "../Option.js"
 import { hasProperty, type Predicate, type Refinement } from "../Predicate.js"
+import type * as Redacted from "../Redacted.js"
 import type * as Secret from "../Secret.js"
 import * as configError from "./configError.js"
 import * as core from "./core.js"
 import * as defaultServices from "./defaultServices.js"
 import * as effectable from "./effectable.js"
 import * as OpCodes from "./opCodes/config.js"
+import * as redacted_ from "./redacted.js"
 import * as InternalSecret from "./secret.js"
 
 const ConfigSymbolKey = "effect/Config"
@@ -417,6 +419,15 @@ export const secret = (name?: string): Config.Config<Secret.Secret> => {
   const config = primitive(
     "a secret property",
     (text) => Either.right(InternalSecret.fromString(text))
+  )
+  return name === undefined ? config : nested(config, name)
+}
+
+/** @internal */
+export const redacted = (name?: string): Config.Config<Redacted.Redacted> => {
+  const config = primitive(
+    "a redacted property",
+    (text) => Either.right(redacted_.make(text))
   )
   return name === undefined ? config : nested(config, name)
 }