+++ packages/sql-mysql2/src/Migrator.ts
@@ -10,7 +10,7 @@ import type { SqlError } from "@effect/sql/Error"
 import * as Migrator from "@effect/sql/Migrator"
 import * as Effect from "effect/Effect"
 import * as Layer from "effect/Layer"
-import * as Secret from "effect/Secret"
+import * as Redacted from "effect/Redacted"
 import { MysqlClient } from "./Client.js"
 
 /**
@@ -52,7 +52,7 @@ export const run: <R2 = never>(
             MYSQL_HOST: sql.config.host,
             MYSQL_TCP_PORT: sql.config.port?.toString(),
             MYSQL_PWD: sql.config.password
-              ? Secret.value(sql.config.password)
+              ? Redacted.value(sql.config.password)
               : undefined
           }),
           Command.string